pyproject.toml
[project]
name = "grok-mcp"
version = "0.1.0"
description = "MCP server for xAI Grok – X/Twitter search and web search"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
grok-mcp = "server:main"
